pyproject.toml
[project]
name = "csv-etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "psycopg[binary,pool]>=3.2.13",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
]
[tool.pytest.ini_options]
pythonpath = ["."]
